requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.1
psycopg2-binary==2.9.9
alembic==1.12.1
openai==1.2.4
python-dotenv==1.0.0
websockets==11.0.3
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
speechrecognition==3.10.0
pydub==0.25.1
boto3==1.28.64
asyncpg==0.28.0 